requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
python-binance==1.0.15
pandas== 1.3.3
numpy==1.21
plotly==5.3.1
ta==0.7.0